Objectifying stats fields [#43]. Implemented a 'fieldFactory' to create stats fields objects, with an aim to simplify and generally improve the code used in storing and fetching data for return. Specifically fields have to toString method which will call its respective format method to ensure all requests for data are returned with the correct number of decimal places, thousands separators, suffixes etc.
I've also updated the integrityCheck method to take advantage of this
refactor and have also included calls to nicer underscore.js
functionality.

Currently the commit should be safe to deploy as commands still return
data held in the old-style structures and new users continue to be
initialized in this fashion also.